requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
altair=4.1.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
astor=0.8.1=pypi_0
async-generator=1.10=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
base58=2.1.1=pypi_0
bleach=4.1.0=pypi_0
blinker=1.5=pypi_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=2.0.12=pypi_0
click=7.1.2=pypi_0
colorama=0.4.5=pypi_0
commonmark=0.9.1=pypi_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pypi_0
faiss=1.5.3=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.18=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
ipykernel=5.5.6=pypi_0
ipython=7.16.3=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.1=pypi_0
jedi=0.17.2=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.1.1=pypi_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.3.7=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=pypi_0
mistune=0.8.4=pypi_0
mmcls=0.24.0=pypi_0
mmcv-full=1.6.1=pypi_0
mmdet=2.25.1=pypi_0
mmsegmentation=0.29.0=dev_0
model-index=0.1.11=pypi_0
nbclient=0.5.9=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=pypi_0
notebook=6.4.10=pypi_0
numpy=1.19.5=pypi_0
opencv-python=4.6.0.66=pypi_0
openmim=0.3.1=pypi_0
openssl=1.1.1q=h7f8727e_0
ordered-set=4.0.2=pypi_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.7.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=pypi_0
pip=21.2.2=py36h06a4308_0
prettytable=2.5.0=pypi_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.30=pypi_0
protobuf=3.19.4=pypi_0
ptyprocess=0.7.0=pypi_0
pyarrow=6.0.1=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pypi_0
pydeck=0.6.2=pypi_0
pygments=2.13.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.0=pypi_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pypi_0
pytz=2022.2.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=23.2.1=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
rich=12.5.1=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py36h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.39.2=h5082296_0
streamlit=0.86.0=pypi_0
tabulate=0.8.10=pypi_0
terminado=0.12.1=pypi_0
terminaltables=3.1.10=pypi_0
testpath=0.6.0=pypi_0
threadpoolctl=3.1.0=pypi_0
timm=0.4.12=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torch=1.9.0+cu111=pypi_0
torchvision=0.10.0=pypi_0
tornado=6.1=pypi_0
tqdm=4.61.2=pypi_0
traitlets=4.3.3=pypi_0
typing-extensions=4.1.1=pypi_0
tzdata=2022.2=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.11=pypi_0
validators=0.20.0=pypi_0
watchdog=2.1.9=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wget=3.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.1=pypi_0
xz=5.2.5=h7f8727e_1
yapf=0.32.0=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.12=h7f8727e_2
